+++ tests/test.py
@@ -7,7 +7,7 @@ handle SIGALRM nopass
 source ~/.gdbinit-gef.py
 """
 
-#@unittest.skip
+@unittest.skip
 class Debugger_process(unittest.TestCase):
 	def setUp(self):
 		warnings.simplefilter("ignore", ResourceWarning)
@@ -108,7 +108,24 @@ class Debugger_process(unittest.TestCase):
 			dbg.debug_from_done.wait()
 			self.assertEqual(dbg.eip, 0x804809d)
 
-#@unittest.skip
+	# Yet impossible. race condition between next's and finish's continue until. TODO handle priority level between step and continue
+	#@unittest.skip
+	def test_multiple_breakpoints(self):
+		print("\ntest_multiple_breakpoints: ", end="")
+		with context.local(arch="amd64", bits=64):
+			dbg = Debugger("./traps_withSymbols", aslr=False)
+			self.debuggers.append(dbg)
+			dbg.until(0x401581)
+			callback_finished = []
+			def callback(dbg):
+				dbg.finish()
+				callback_finished.append(dbg.instruction_pointer)
+				log.info("address added")
+				return False
+			dbg.b(0x433494, callback=callback)
+			dbg.next()		
+			self.assertEqual(callback_finished[0], 0x401586)
+@unittest.skip
 class Debugger_actions(unittest.TestCase):
 	def setUp(self):
 		warnings.simplefilter("ignore", ResourceWarning)
@@ -119,12 +136,12 @@ class Debugger_actions(unittest.TestCase):
 		self.dbg.close()
 
 	# Fail, ho fatto next a mano e ho perso il controllo
-	@unittest.skip
+	#@unittest.skip
 	def test_continue_until(self):
 		print("\ntest_continue_until: ", end="")
 		self.dbg.b(0x403ad7, temporary=True)
 		print("\nPlay with gdb once we hit address 0x403ad7 and then send continue")
-		self.dbg.c(until=0x403adb)
+		self.dbg.continue_until(0x403adb)
 		self.assertEqual(self.dbg.instruction_pointer, 0x403adb)
 
 	#@unittest.skip
@@ -135,7 +152,7 @@ class Debugger_actions(unittest.TestCase):
 		done.wait()
 		self.assertEqual(self.dbg.instruction_pointer, 0x4038c2)
 
-#@unittest.skip
+@unittest.skip
 class Debugger_callbacks(unittest.TestCase):
 	def setUp(self):
 		warnings.simplefilter("ignore", ResourceWarning)
@@ -173,8 +190,10 @@ class Debugger_callbacks(unittest.TestCase):
 		self.dbg.c(wait=True)
 		#self.dbg.interactive()
 		self.assertEqual(self.dbg.rip, 0x403ad9)
-		self.assertFalse(len(self.dbg.breakpoints))
+		self.assertFalse(len(self.dbg.breakpoints[0x403ad0] + self.dbg.breakpoints[0x403ad9]))
+		self.assertFalse(self.dbg.priority)
 
+	# It works, but it should finish with priority 0, not 1
 	#@unittest.skip
 	def test_step(self):
 		print("\ntest_step: ", end="")
@@ -186,15 +205,28 @@ class Debugger_callbacks(unittest.TestCase):
 		self.dbg.b(0x403ad0, callback=callback, temporary=True)
 		self.dbg.c(wait=True)
 		self.assertEqual(self.dbg.rip, 0x403ad2)
+		self.assertFalse(self.dbg.priority)
 
-	# Ricordati di testare anche finish con callback
+	#@unittest.skip
+	def test_enforce_stop(self):
+		print("\ntest_enforce_stop: ", end="")
 
+		def callback(dbg):
+			dbg.finish()
+			return False
 
-#@unittest.skip
+		self.dbg.b(0x400870, callback=callback)
+		#self.dbg.b(0x40388e)
+		#self.dbg.c(wait=True)
+		self.dbg.until(0x40388e)
+		self.assertEqual(self.dbg.rip, 0x40388e)
+		self.assertFalse(self.dbg.priority)
+
+@unittest.skip
 class Debugger_memory(unittest.TestCase):
 	def setUp(self):
 		warnings.simplefilter("ignore", ResourceWarning)
-		self.dbg = Debugger("./cube")
+		self.dbg = Debugger("./cube", aslr=False, from_start=False)
 
 	def tearDown(self):
 		if hasattr(self, "dbg"):
@@ -211,6 +243,7 @@ class Debugger_memory(unittest.TestCase):
 		self.assertEqual(self.dbg.rax, 0xdeadbeeffafa90be)
 		self.dbg.r11 = 0x134343432342
 		self.assertEqual(self.dbg.r11, 0x134343432342)
+		self.assertFalse(self.dbg.priority)
 
 	#@unittest.skip
 	def test_special_registers(self):
@@ -218,13 +251,12 @@ class Debugger_memory(unittest.TestCase):
 		self.assertEqual(self.dbg.return_value, self.dbg.rax)
 		self.assertEqual(self.dbg.stack_pointer, self.dbg.rsp)
 		self.assertEqual(self.dbg.instruction_pointer, self.dbg.rip)
+		self.assertFalse(self.dbg.priority)
 
 	# TODO: Test it with multiple inferiors if possible
 	#@unittest.skip
 	def test_alloc(self):
 		print("\ntest_alloc: ", end="")
-		self.dbg.close()
-		self.dbg = Debugger("./cube", aslr=False, from_start=False) # Wait for the libc to be loaded
 		pointer = self.dbg.alloc(16)
 		self.dbg.write(pointer, p64(0xdeadbeeffafa90be))
 		self.assertTrue(hex(pointer) in self.dbg.execute("heap chunks")) # WARNING THIS ONLY WORKS WITH GEF
@@ -233,9 +265,31 @@ class Debugger_memory(unittest.TestCase):
 		self.dbg.write(pointer, p64(0xdeadbeeffafa90be))
 		self.dbg.dealloc(pointer, len=16, heap=False)
 		# remember dealloc in the bss will only delete the last chunk... 
+		self.assertFalse(self.dbg.priority)
 		self.dbg.close()
-		
-#@unittest.skip
+
+	#@unittest.skip
+	def test_writes(self):
+		print("\ntest_writes: ", end="")
+		ints = [4, 6, 295342, 23, 50032]
+		strings = [b"ciao", b"come stai ?", b"questo programma fa davvero schifo"]
+		pointer = self.dbg.write_ints(None, ints)
+		self.assertEqual(ints, self.dbg.read_ints(pointer, len(ints)))
+		self.dbg.write_ints(pointer, [x*2 for x in ints])
+		self.assertEqual([x*2 for x in ints], self.dbg.read_ints(pointer, len(ints)))
+		pointer = self.dbg.write_longs(None, ints)
+		self.assertEqual(ints, self.dbg.read_longs(pointer, len(ints)))
+		self.dbg.write_longs(pointer, [x*2 for x in ints])
+		self.assertEqual([x*2 for x in ints], self.dbg.read_longs(pointer, len(ints)))
+		pointer = self.dbg.write_ints(None, ints[0])
+		self.assertEqual(ints[0], self.dbg.read_ints(pointer, 1))
+		pointer = self.dbg.write_strings(None, strings*100)
+		self.assertEqual(strings*100, self.dbg.read_strings(pointer, len(strings)*100))
+		pointer = self.dbg.write_strings(None, strings[0])
+		self.assertEqual(strings[0], self.dbg.read_strings(pointer, 1))		
+		self.assertFalse(self.dbg.priority)
+
+@unittest.skip
 class Debbuger_fork(unittest.TestCase):
 	from base64 import b64encode
 	def setUp(self):
@@ -267,19 +321,24 @@ class Debbuger_fork(unittest.TestCase):
 		"""
 		self.dbg = Debugger("./httpd", aslr=False, script=gdbscript)
 		CALL_TO_B64DECODE = 0x26d5
-		self.dbg.b(CALL_TO_B64DECODE)
-		self.dbg.c(wait=False)
+		#self.dbg.b(CALL_TO_B64DECODE)
+		#self.dbg.c(wait=False)
+		done = self.dbg.until(CALL_TO_B64DECODE, wait=False)
 		self.dbg.p.sendline(self.http_request(keepAlive=True))
-		self.dbg.wait()
+		#self.dbg.wait()
+		done.wait()
 		self.assertEqual(self.dbg.rip, 0x5555555566d5)
 		self.dbg.c(wait=True)
 		self.dbg.execute("inferior 1") # We can't go back while the child is running
-		self.dbg.c(wait=False)
+		#self.dbg.c(wait=False)
+		done = self.dbg.until(CALL_TO_B64DECODE, wait=False)
 		self.dbg.p.sendline(self.http_request(keepAlive=True))
-		self.dbg.wait()
+		#self.dbg.wait()
+		done.wait()
 		self.assertEqual(self.dbg.rip, 0x5555555566d5)
+		self.assertFalse(self.dbg.priority)
 		self.dbg.close()
-	
+		
 	#@unittest.skip
 	def test_split(self):
 		print("\ntest_split: ", end="")
@@ -304,6 +363,7 @@ class Debbuger_fork(unittest.TestCase):
 			dbg.p.sendline(self.http_request(keepAlive=True))
 			dbg.c(wait=True)
 			self.assertEqual(dbg.rip, 0x5555555566d5)
+			self.assertFalse(dbg.priority)
 			dbg.close()
 			child.close()
 
@@ -315,12 +375,13 @@ class Debbuger_fork(unittest.TestCase):
 			dbg.c(wait=False)
 			pid = dbg.wait_split() # and then for the child to split out
 			child = dbg.children[pid]
+			sleep(0.05) # Wait for the priority to reach 0
 			self.assertEqual(dbg.instruction_pointer, 0x4327a7) # Will continue without interuptions. Get's there waiting for the PTRACE from the child
 			self.assertEqual(child.instruction_pointer, 0x432d37)
+			self.assertFalse(dbg.priority)
 			dbg.close()
 			child.close()
 
-	# This one fails 1/4 times
 	#@unittest.skip
 	def test_ptrace_emulation(self):
 		print("\ntest_ptrace_emulation: ", end="")
@@ -332,33 +393,36 @@ class Debbuger_fork(unittest.TestCase):
 			dbg = Debugger("./traps_withSymbols", script=gdbinit, aslr=False, debug_from=ANTI_DEBUG_TEST_FINISHED).set_split_on_fork()
 			
 			dbg.continue_until("fork")
+			# si blocca se non ha bisogno di mandare l'interrupt [19/06/23]
 			dbg.finish()
 			pid = dbg.wait_split()
 			
 			second_child = dbg.children[pid]
-			second_child.emulate_ptrace_slave(dbg)
-			dbg.emulate_ptrace_master(second_child)
+			second_child.emulate_ptrace()
+			dbg.emulate_ptrace(manual=True) # Stop on waitpid
 			# Continue after fork
-			dbg.c(wait=True)
 			second_child.c(wait=True)
+			dbg.c(wait=True)
 		
 			# handle signal
 			dbg.c(wait=True)
 			second_child.c(wait=True, force=True)
-			
+			log.info("setup done")
 			dbg.p.sendline(b"CSCG{4ND_4LL_0FF_TH1S_W0RK_JU5T_T0_G3T_TH1S_STUUUP1D_FL44G??!!1}")
 			for i in range(1, 22):
 				if second_child.instruction_pointer == RWX_SECTION + 1:
+					# reach waitpid
+					dbg.c(wait=True)
 					# setup unpack
 					dbg.c(wait=True)
-					second_child.c(until=END_UNPACK)
+					second_child.until(END_UNPACK, loop=True)
 					if i < 4:
 						second_child.c(wait=True)
 						continue
 					elif i == 4:
 						dbg.p.recv() # Just receive the prompt
 						# Pass ptrace check
-						second_child.c(until=SYSCALL_TRAP_PTRACE)
+						second_child.until(SYSCALL_TRAP_PTRACE, loop=True)
 						second_child.step()
 						second_child.return_value = 0x0
 					else:
@@ -369,11 +433,12 @@ class Debbuger_fork(unittest.TestCase):
 				second_child.c(wait=True)
 			
 			self.assertTrue(b"YES !" in dbg.p.recv())
+			self.assertFalse(dbg.priority)
 
 			second_child.close()
 			dbg.close()
 
-#@unittest.skip
+@unittest.skip
 class Debugger_signals(unittest.TestCase):
 	def setUp(self):
 		warnings.simplefilter("ignore", ResourceWarning)
@@ -393,7 +458,7 @@ class Debugger_signals(unittest.TestCase):
 		HANDLER_RET = 0x04011ff
 		CHECK_CALL = 0x0401341
 		self.dbg.p.sendline(b"serial_a_caso")
-		self.dbg.cont(until=CHECK_CALL)
+		self.dbg.until(CHECK_CALL)
 		self.dbg.step()
 		self.dbg.next_signal = False
 		output = []
@@ -412,8 +477,9 @@ class Debugger_signals(unittest.TestCase):
 		self.dbg.step_until_ret(callback)
 		with open("dump_ExceptionalChecking") as fp:
 			self.assertEqual(output, fp.read().split("\n"))
+		self.assertFalse(self.dbg.priority)
 
-#@unittest.skip
+@unittest.skip
 class Debugger_calls(unittest.TestCase):
 	def setUp(self):
 		warnings.simplefilter("ignore", ResourceWarning)
@@ -430,16 +496,18 @@ class Debugger_calls(unittest.TestCase):
 		for mossa in [0x00003175, 0x00003e74, 0x000038d9, 0x00001885]:
 			# Create blank cube
 			pointer = self.dbg.alloc(54*8)
-			for i in range(54):
-				self.dbg.write(pointer+i*8, p64(i))
+			self.dbg.write_longs(pointer, list(range(54)))
+			#for i in range(54):
+			#	self.dbg.write(pointer+i*8, p64(i))
 
 			self.dbg.call(mossa, [pointer])
 			out.append(f"prossima funzione {hex(self.dbg.get_base_elf() + mossa)}")
-			for i in range(54):
-				out.append(f"{i}: {u64(self.dbg.read(pointer+i*8, 8))}")
+			for i, value in enumerate(self.dbg.read_longs(pointer, 54)):
+				out.append(f"{i}: {value}")
 		#print("\n\n".join(out)) 
 		with open("./dump_Cube") as fd:
 			self.assertEqual(out, fd.read().split("\n"))
+		self.assertFalse(self.dbg.priority)
 		self.dbg.close()
 
 	#@unittest.skip
@@ -456,6 +524,7 @@ class Debugger_calls(unittest.TestCase):
 			self.dbg.syscall(constants.SYS_close, [fd])
 			with open(path, "rb") as file:
 				self.assertEqual(file.read(), data)
+			self.assertFalse(self.dbg.priority)
 
 #class Debugger_fancy_gdb(unittest.TestCase):
 #
@@ -485,7 +554,7 @@ class Debugger_calls(unittest.TestCase):
 #		self.dbg.write(pointer, p64(0xdeadbeeffafa90be))
 #		self.dbg.close()
 
-@unittest.skip
+#@unittest.skip
 class Debugger_libdebug(unittest.TestCase):
 	def setUp(self):
 		warnings.simplefilter("ignore", ResourceWarning)
@@ -493,7 +562,7 @@ class Debugger_libdebug(unittest.TestCase):
 	def tearDown(self):
 		self.dbg.close()
 
-	#@unittest.skip
+	@unittest.skip
 	def test_migrate(self):
 		print("\ntest_migrate: ", end="")
 		with context.local(arch="amd64", bits=64):
@@ -506,8 +575,9 @@ class Debugger_libdebug(unittest.TestCase):
 			self.assertEqual(self.dbg.instruction_pointer, 0x7ffff7fe32b3)
 			self.dbg.step()
 			self.assertEqual(self.dbg.instruction_pointer, 0x7ffff7fe4050)
+			self.assertFalse(self.dbg.priority)
 
-	#@unittest.skip
+	@unittest.skip
 	def test_continue_until(self):
 		print("\ntest_continue_until [libdebug]: ", end="")
 		with context.local(arch="amd64", bits=64):
@@ -515,6 +585,7 @@ class Debugger_libdebug(unittest.TestCase):
 			self.dbg.migrate(libdebug=True)
 			self.dbg.continue_until("main")
 			self.assertEqual(self.dbg.instruction_pointer, 0x55555555570c)
+			self.assertFalse(self.dbg.priority)
 
 	@unittest.skip
 	def test_call(self):
@@ -525,8 +596,9 @@ class Debugger_libdebug(unittest.TestCase):
 			self.dbg.migrate(libdebug=True)
 			address = self.dbg.call("malloc", [0x100])
 			self.assertEqual(address, 0x55555555d2a0)
+			self.assertFalse(self.dbg.priority)
 			
-	#@unittest.skip
+	@unittest.skip
 	def test_callbacks(self):
 		print("\ntest_callbacks [libdebug]: ", end="")
 		with context.local(arch = 'amd64'):
@@ -542,11 +614,11 @@ class Debugger_libdebug(unittest.TestCase):
 			
 			second_child = self.dbg.children[pid]
 			self.dbg.migrate(libdebug=True)
-			second_child.emulate_ptrace_slave(self.dbg)
-			self.dbg.emulate_ptrace_master(second_child)
+			second_child.emulate_ptrace()
+			self.dbg.emulate_ptrace()
 			# Continue after fork
-			self.dbg.c(wait=True)
 			second_child.c(wait=True)
+			self.dbg.c(wait=True)
 			return
 			# handle signal
 			self.dbg.c(wait=True)
@@ -555,6 +627,7 @@ class Debugger_libdebug(unittest.TestCase):
 			for i in range(1, 22):
 				if second_child.instruction_pointer == RWX_SECTION + 1:
 					# setup unpack
+					self.dbg.c(wait=True)
 					self.dbg.c(wait=True)
 					second_child.c(until=END_UNPACK)
 					if i < 4:
@@ -574,6 +647,7 @@ class Debugger_libdebug(unittest.TestCase):
 				second_child.c(wait=True)
 			
 			self.assertTrue(b"YES !" in dbg.p.recv())
+			self.assertFalse(dbg.priority)
 
 			second_child.close()
 
@@ -628,6 +702,7 @@ class Debugger_libdebug(unittest.TestCase):
 		    flag += p32(n)
 		flag = xor(flag, 0xd)
 		self.assertEqual(flag, b"CSCG{4ND_4LL_0FF_TH1S_W0RK_JU5T_T0_G3T_TH1S_STUUUP1D_FL44G??!!1}")
+		self.assertFalse(self.dbg.priority)
 
 if __name__ == "__main__":
 	with context.quiet: